+++ packages/react-ui/src/components/atoms/LabelWithIndicator.js
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import { Box, styled } from '@mui/material';
 
 import Typography from './Typography';
+import { ICON_SIZE_SMALL } from '../../theme/themeConstants';
 
 const Root = styled(Box)(({ theme }) => ({
   display: 'flex',
@@ -16,30 +17,52 @@ const LabelIndicator = styled(Typography)(({ theme }) => ({
   }
 }));
 
-const LabelWithIndicator = ({ label, type }) => {
+const Icon = styled(Box)(({ theme }) => ({
+  display: 'flex',
+
+  svg: {
+    width: ICON_SIZE_SMALL,
+    height: ICON_SIZE_SMALL,
+    fontSize: ICON_SIZE_SMALL,
+
+    path: {
+      fill: theme.palette.text.secondary,
+
+      '.Mui-disabled &': {
+        fill: theme.palette.text.disabled
+      }
+    }
+  }
+}));
+
+const LabelWithIndicator = ({ label, type, icon, inheritSize }) => {
   const isRequired = type === 'required';
 
   return (
     <Root>
       {label}
-      <LabelIndicator
-        component='span'
-        variant='inherit'
-        color='textSecondary'
-        weight='regular'
-      >
-        {isRequired ? '(required)' : '(optional)'}
-      </LabelIndicator>
+
+      {type && (
+        <LabelIndicator
+          component='span'
+          variant={inheritSize ? 'inherit' : 'caption'}
+          color='textSecondary'
+          weight='regular'
+        >
+          {isRequired ? '(required)' : '(optional)'}
+        </LabelIndicator>
+      )}
+
+      {icon && <Icon>{icon}</Icon>}
     </Root>
   );
 };
 
-LabelWithIndicator.defaultProps = {
-  type: 'optional'
-};
 LabelWithIndicator.propTypes = {
   label: PropTypes.oneOfType([PropTypes.element, PropTypes.string]).isRequired,
-  type: PropTypes.oneOf(['optional' | 'required'])
+  type: PropTypes.oneOf(['optional', 'required']),
+  icon: PropTypes.element,
+  inheritSize: PropTypes.bool
 };
 
 export default LabelWithIndicator;