+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 
+- LabelWithIndicator: UI improvements [#819](https://github.com/CartoDB/carto-react/pull/819)
 - SelectField: extend to accept more data structures as children and to fix placeholder [#796](https://github.com/CartoDB/carto-react/pull/796)
 
 ## 2.3